test: Add FakeAuthenticationServiceImpl for tests
viewModel in StrenApp is of type StrenAppViewModel and injected using
hiltViewModel(), so creating and injecting fake ViewModel class will be
 tedious.

Instead, a FakeAuthenticationServiceImpl that's injected into the
viewModel using constructor and factory is more simple.

The downside is that some composables lower in the hierarchy may use
AuthenticationServiceImpl so it's currently impossible to test some
navigation flows that involve them.