fix: Screens overlapping in TrainingTabsScreen
Description: The screens are supposed to be shown one at time
when the correct tab is selected. However, all screens are shown at
the same time overlapping each other

Cause: The screen composables passed to TrainingTabsScreen as parameters
are of type "Unit", which is not a function type. So, when TrainingGraph
passes these composables to TrainingTabsScreen as parameters, it
accidentally invoked these composables, causing all of screens to be
composed immediately outside of the TrainingTabsScreen

Solution: Change the parameter type from "Unit" to
"@Composable() -> Unit" which is a function type. Then only invoke the
composables inside TrainingTabsScreen after setting up TabRow and
HorizontalPager