feat: Let composables configure app bar
Description:
- Data class AppBarConfiguration represents all the information an
app bar needs to display itself (icons, title, actions...)
- A lambda is passed down the composable hierarchy for each screen
to configure app bar

Note:
- The original idea was to make AppBarConfiguration an enum class
and add the logic for deciding which configuration to use based on
current destination.
- However, this was abandoned as each composable has its own actions
which can only be defined in the composable function itself (Ex: actions
that depend on ViewModel, so a pre-defined enum class is not sufficient)
- Also, for composables that use TabLayout, switching between tabs
doesn't use NavController (since that will erase the TabRow), so the
tab screens won't be registered in the backstack. This will render the
original idea undoable as the core logic for changing
AppBarConfiguration is based on current destination of backstack.

TODO: The chosen method is quite arduous as the lambda for configuring
AppBar must be passed to all composables in the hierarchy, any new one
added in the future must also do the same . Or else, the app bar won't
change back to default even if the next destination doesn't have any
special need for the app bar.