refactor: Replace TitleResId with String in AppBar
The "titleResourceId" field of type Int in AppBarConfiguration allows
this field to have a non-hardcoded default value by setting a resouce id
as the default value, and screens will customize this by passing a
string resource id as parameter.

However, this is not suitable for cases when app bar title is dynamic
and can't be statically stored in string resources, so the field is
changed to "title" of type String. The downside of this approach is you
can't rely on string resources for default value (accessing string
resources require being in a composable function) and have to use
a hardcoded string as default value.