feat: Implement routine detail TextFields
TextFields composable are usually accompanied by a MutableState variable
to show TextField's value and a lambda to change that value, however, in
this screen, the number of TextFields could reach 20-30 on average (in
theory, although unrealistic, this number can be infinite) so the above
solution may bloat the UI code

The chosen solution is to pass a lambda from ViewModel down the
composable hierarchy and let ViewModel takes care of all the value
update of the TextFields. Currently, this solution has a drawback of
potentially affecting performance. Since AddEditRoutineViewModel only
exposes a single source of data for all of the components in
AddEditRoutineScreen to use, any small change to any of the TextField
will lead to change in this single data source, which can lead to
recomposition of every UI component that depends on it