fix: App crash when holding delete button
Description: WHen user holds the delete button on trainingSet TextFields
in AddEditRoutineScreen to delete the whole input, app crashes

Cause: Value exposed by AddEditRoutineViewModel is immutable, so every
update to its data requires creating new objects. Basically 2 data flows
happen asynchronously:
- [Read data] AddEditRoutineViewModel exposes objects ->
AddEditRoutineScreen reads objects
- [Write data] AddEditRoutineViewModel tries to update
AddEditRoutineViewModel -> New objects are created in
AddEditRoutineViewModel to update

The app crashes come from the fact that 2nd flow outruns the 1st flow,
leading the following situation:

1. [Read] ViewModel sends object A to Screen
2. [Write] Screen sends update requests to object A - ViewModel creates
object B and deletes object A to update itself (object creation is
required since data is immutable in ViewModel)
3. [Write] Screen sends even more update requests to object A, since
the writing rate is too high, Screen hasn't had the chance to read
object B, and still thinks it needs to updates object A
4. [Write] ViewModel tries to create object C to replace object A, but
object A is already deleted in Step 2., it's just that Screen writes
too quickly and didn't receive the news that object A is already
replaced by object B
5. Object A no longer exists in ViewModel so ViewModel is trying to
reach a deleted object and app crashes

Options:
1. Updates to the TextFields shouldn't update the data source in
ViewModel, instead, create separate data sources for each (MutableState
for TextFields, StateFlow for everything else) and only merge the 2 when
the user saves
2. Instead of updating based on object identity like original solution,
update based on the object's exerciseId field, which will live through
object creation. However, this removes the possibility of having the
same exercise at different order in the routine (Ex: Exercise A ->
Exercise B -> Exercise A), since updates will be sent to all exercises
with the  same exercise id (to prevent this was the reason why
updates used object identity in the original solution)
3. (Currently) Add "id" field (which is a random UUID) to the data that
needs to be updated. This is a combination of the original solution
with the 2nd one: Allow for the same exercise appear separately while
living through object creation. However, the drawback of this option
is that when user deletes TextField too fast, instead of app crash, some
old characters are not deleted fast enough and stay in the TextField