fix: Food data returned by API has diff schemas
Problem: The FDC API endpoints for getting food and searching food by
name return data of different schemas for the "foodNutrients" field
(which is strange, since the API Specs indicates that they should be
the same). Since the application designs model classes based
on the schema described by FDC API specs, the "foodNutrients" field
is converted to the NetworkFoodNutrient class whether application is
simply fetching food or searching food

Problem 2: This problem can be solved by adding another
NetworkFoodNutrient model class to handle both cases, but it'll require
changing a lot of classes to use the new model class

Solution: Turn the old model class into a sealed interface with
fields for child classes to override (abstract class is actually a more
intuitive structure, but will lead to serialization problem). Then,
each child class of this interface will deal with a different schema.

This solution works since application needs roughly the same
information for each "foodNutrient" object (name, amount, unit)
regardless of foodNutrient's API schema, and the root cause of the
problem only is that each schema has different names for these info. If
in the future, FDC API removes  any of the above info from any of the 2
schemas, the application will break

The final step is to generify NetworkFood, the class that contains a
List of NetworkFoodNutrient object (The application actually
deserializes API data to this class, and recursively deserializes
foodNutrients field to NetworkFoodNutrient). The generic class
NetworkFood then helps application decide which model
class to actually deserialize to (since NetworkFoodNutrient is just an
interface, it doesn't have any serialization info, only its child
class does)