feat: Add a new nav arg for EditFoodEntryScreen
The new nav arg is the name of meal being edited. This is for the case
where user is editing meal that exists in the application, but not
on database (The default meals - Breakfast, Lunch, Dinner, Snacks - are
not added to the database until user edits one), so the meal id nav arg
is not sufficient for editing (Since there's no data in database with
that id yet to fetch meal data)

With the meal id and meal name nav arg, which are 2/3 of fields of
the Meal class (the last of is list of foods, which will be constructed
using the current EditFoodEntryScreen), the application will
actually add a new whole meal to the database (although it will look
like the application is editing an existing meal to the user). And
this addition can only be done if all the info about the meal is
available, hence the new nav arg

TODO: This is actually very brittle and requires modifications if in
the future, more fields are added to the Meal class (since there will
be more nav args to add). A better solution would be to create
a table/collection consisting solely of the default meals on the
database. Then all screens can depend on this collection to fetch
meal data, instead of creating the default meals in the application.

Another solution is adding default meals to all dates on the database,
which is overkill and arduous to implement/maintain