feat: Remove DataFetchingTrigger
The old data flow for DashboardViewModel and DashboardScreen was
to collect a flow from GetUserFullDataUseCase, then updates
DataFetchingTrigger in DashboardViewModel, from which DataOutput for
DashboardScreen is created.

DataFetchingTrigger basically acted as a local cache that allows for
localizing changes, which means that small changes (Ex: Changing a
tracked category's dates range) won't reset the whole screen. However,
this brings up the issue of cache invalidation. Since
DataFetchingTrigger relies on MutableStateFlow - a data holder whose
only way to update is to create a whole new object, cache invalidation
is simply removing all of the old data with the new one, which will
reset the whole screen anyway (the problem we try to solve with cache
in the first place)

The alternative is to collect straight from the GetUserFullDataUseCase
and directly create DataOutput without using a cached. Firestore's
snapshots() method that returns a Flow with
backend-real-time-updates allows for this