feat: Modify methods of AuthenticationService
Including:
- Add getUserAuthStateStream() which returns an auto-updated Flow
of user auth state. The addAuthStateListeners() which was the callback
version of this method worked well for plain callbacks, but falls short
when developers need to call suspend functions in these callbacks.
getUserAuthStateStream() returns a Flow so the above shortcoming doesn't
exist
- Modify authenticate() method (which basically signs user in) so it
only succeeds if the email being authenticated is already verified by
user, or else an exception is thrown for the UI layer to handle.
Without the above logic, Firestore will just signs users in without
a care for whether users have their emails verified or just use
some random emails
- Add signOut() method for signing out