feat: Keep SplashScreen on until app navigates
Before: LoginScreen is shown automatically whether user has signed in
or not, then when the auth state (user signed in/not signed in) is
evaluated, application navigates to the right screen. This is quite
clunky since users that have already signed in are forced to see the
LoginScreen

Now: The SplashScreen is kept on until application navigates to the
right screen

Note: This was implemented by calling delay() for an amount of time,
which basically buys time for the application to do its navigation,
which is too brittle. If on some devices, the navigation takes more
time, this implementation becomes useless. On the other hand, if the
navigation takes less time, then user is forced to start at the
SplashScreen for no reason