fix: Faulty logic for "shouldShowTopBar"
"shouldShowTopBar" is a field in StrenAppState for deciding when to
show the top app bar. Since the top bar should be shown everywhere
but the authentication-related screens, the old logic was fine. Now
that onboarding-related screens on which the top app bar should be
hidden are added, the old logic is not sufficient