fix: App shows Auth screens even if user signed in
Cause: StrenNavHost which contains all screen composables, observes
the auth state and navigates to the correct screen based on auth state
change. However, this takes too long, so the starting route (which is
Login screen) is always shown before application navigates to the
correct screen

Solution: The auth state in StrenNavHost is monitored through a Boolean
var called "isUserSignedIn" which can only indicates whether user's
signed in or not. The solution is to make this var nullable, which means
there are 3 possibilities:
- isUserSignedIn = null: Auth state not decided yet
- isUserSignedIn = false: Auth state captured and user hasn't signed in
- isUserSignedIn = true: Auth state captured and user's signed in

While "isUserSignedIn" is null, keep the splashscreen on and stall for
time until the auth state is decided and the correct screen is shown