fix: User not navigated after login with Facebook
Cause: Due to Firestore's implementation, user will automatically
be signed in after they signs up with email, regardless of whether
that email is verified or not. To counter this, application will
evaluate whether the "supposedly signed in" user has their email
verified or not, and only navigate to Onboarding/Dashboard if the above
is true.

The issue occurs because when user logs in with Facebook, their email
is still not marked as verified by Firestore, although login with Gmail
is fine.

Solution: The workaround is to add a check whether the signed in user
is using Facebook as their authentication provider, and not check
for email verification state if the above is true

Reference:
https://stackoverflow.com/questions/38398545/firebase-facebook-auth-email-verified-always-false
https://github.com/firebase/firebase-js-sdk/issues/340