fix: Duplicated goals in NutritionDiaryScreen
Core nutrient goals for a day in NutritionDiaryScreen are constructed by
(1) fetching default core nutrients (represented by an enum) then (2)
search for these nutrients by "nutrientName" field in the current day's
foods.

Cause: The (2) step was done by using "name" field of the enum from
step (1) instead of the "nutrientName" field. This is because each enum
value automatically has a "name" field that represents the NAME of the
enum value itself, but what we need is a field called "nutrientName"
that each enum value must define by hand

Solution: Stop using the "name" field. Another solution is to rename
the enum values which might violate naming conventions. Overriding
"name" field was also considered but this is prohibited in Kotlin