fix: DashboardScreen not receive real-time updates
Description: When user tracks new category, application is supposed to
show the newly tracked category on DashboardScreen but it didn't

Cause: When user tracks new category,  a new element is added to
Firestore document (1). Then, the snapshots() method from Firestore will
expose an automatically updated flow of the newest elements (2). But,
for some reason, if the (1) was done the usual way (calling Firestore's
update()), (2) won't be updated even if Firestore has its new elements

Solution: Have to do (1) by running a Firestore transaction ??