refactor: Simplify nav args for TrainingHistory
The old ways of handling nav args was to put the keys defining each
screen's nav args in the composable file of the screen's route itself.
However, these nav arg keys are too long, and are a pain when some
similar  screens in a small nested graph have nav args with the same
name, which leads to a dilemma: either adds some more text to the nav
arg key to  tell the nav args apart for each screen (the name is now
longer) or use  full-blown qualifier (package name + nav arg name) for
these names (which is even longer than the 1st option)

The new way is to hardcode the nav args name of each route in the file
that contains the nested graph of these routes (In this case,
TrainingHistoryNavigation.kt), which improves readability. This is
acceptable due to 2 reasons:
- The nav args names are only accessed in the nested graph file,
so any bugs related to mismatched nav arg names are can only occur
within this file. So it's worth trading non-duplicated nav arg names
for readability
- For the screens with similar nav args, as long as their screen routes
are different, it's enough to identify these screens. So, having
nav args with the same name won't be a problem