feat: Implement WorkoutInProgressViewModel
A workout in progress is different from a logged workout. A logged
workout is simply a workout whose details are noted down in the
training history (which means the date this workout is logged can be
different from the date it's actually done).

On the other hand, a workout in progress is a workout that's actually
happening in real time, and the user needs some extra functions (rest
timers, checkboxes to mark which training sets are done, notification on
screen that a workout is  occurring). Therefore, a workout in progress
is much more complicated and involves multiple screens.

This class holds the info of the current workout in progress (training
sets, timers) and expose the same info for any class in the composition
hierarchy to access. This ways, it acts as a singleton of the
workout in progress