fix: Workout refreshed when back from other screen
Expect: When user starts a workout, they're navigated to
StartWorkoutScreen. Then even when they presses back, the workout is
still in progress and a Card is shown on the screen to indicate this.
User can press this card and go back to StartWorkoutScreen with all the
input data intact.

Bug: The data is refreshed when user goes back to StartWorkoutScreen
from other screens

Cause: StartWorkoutScreen takes workout in progress data from
WorkoutInProgressViewModel which requires some args. Since
WorkoutInProgressViewModel is scoped to MainActivity which exists before
any composables, it's impossible to pass these args with navigation. So
args passing is done through a setter method in
WorkoutInProgressViewModel. However, since StartWorkoutScreen is a
composable, whose recompositions happen all the time, the args in
WorkoutInProgressViewModel are set multiple times

Solution: Check if the args in WorkoutInProgressViewModel are
initialized, and only set these args if they are not initialized yet