feat: Show WorkoutOptionDialog
When user is viewing today's workout history, they can either start or
log workout, while if they're viewing any other day, they can only
log workout.

When the history is empty, the above actions are done through buttons at
the bottom of the screen. On the other hand, when the history is not
empty, these are done through an app bar button. However, this button
was implemented to only work for logging workout

This commit implements the app bar button to show a dialog for choosing
between logging or starting a workout if the selected date is the
current calendar day